ci: split integration and library tests, run int only on linux
