+++ .github/workflows/ci.yaml
@@ -6,7 +6,7 @@ jobs:
   lint:
     name: Lint
     runs-on: ubuntu-latest
-    continue-on-error: true
+    fail-fast: false
     steps:
       - name: Checkout the repository
         uses: actions/checkout@v3
@@ -18,15 +18,13 @@ jobs:
         run: |
           npm ci
           npm run lint
-  test:
-    name: Test
+  test-library:
+    name: Test (Library)
+    fail-fast: false
     strategy:
       matrix:
         node-version: [lts/*, latest]
         os: [ubuntu-latest, windows-latest, macos-latest]
-        include:
-          - node-version: 16
-            os: ubuntu-latest
     runs-on: ${{ matrix.os }}
     steps:
       - name: Checkout the repository
@@ -39,15 +37,34 @@ jobs:
         run: |
           npm ci
           npm run test:lib
+  test-integration:
+    name: Test (Integration)
+    fail-fast: false
+    strategy:
+      matrix:
+        node-version: [lts/*, latest]
+        os: [ubuntu-latest]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Checkout the repository
+        uses: actions/checkout@v3
+      - name: Use Node ${{ matrix.node-version }}
+        uses: actions/setup-node@v3
+        with:
+          node-version: ${{ matrix.node-version }}
+      - name: Run integration tests
+        run: |
+          npm ci
+          npm run test:int
   publish:
     name: Publish
-    needs: [lint, test]
+    needs: [lint, test-library, test-integration]
     if: startsWith(github.ref, 'refs/tags/v')
     runs-on: ubuntu-latest
     steps:
       - name: Checkout the repository
-        uses: actions/checkout@v2
-      - uses: actions/setup-node@v1
+        uses: actions/checkout@v3
+      - uses: actions/setup-node@v3
         with:
           node-version: lts/*
           registry-url: https://registry.npmjs.org/
