+++ jest.config.json
@@ -3,7 +3,7 @@
 	"collectCoverage": true,
 	"collectCoverageFrom": ["source/**/*.ts"],
 	"testTimeout": 30000,
-	"testMatch": ["**/test/**/*-test.[jt]s?(x)"],
+	"testMatch": ["**/test/{library,integration}/*-test.[jt]s?(x)"],
 	"moduleFileExtensions": ["js", "jsx", "json", "ts", "tsx"],
 	"moduleNameMapper": {
 		"^(\\.{1,2}/.*)\\.js$": "$1"
