+++ test/library/middleware-test.ts
@@ -1,11 +1,11 @@
-// /test/middleware-test.ts
+// /test/library/middleware-test.ts
 // Tests the middleware by passing in various options to it
 
 import EventEmitter from 'node:events'
 import { jest } from '@jest/globals'
-import slowDown from '../source/index.js'
-import { expectDelay, expectNoDelay } from './helpers/requests.js'
-import { MockStore } from './helpers/mock-stores.js'
+import slowDown from '../../source/index.js'
+import { expectDelay, expectNoDelay } from '../helpers/requests.js'
+import { MockStore } from '../helpers/mock-stores.js'
 
 describe('middleware behaviour', () => {
 	beforeEach(() => {
