+++ test/library/options-test.ts
@@ -1,7 +1,7 @@
-// /test/options-test.ts
+// /test/library/options-test.ts
 // Tests the parsing/handling of options passed in by the user
 
-import slowDown from '../source/index.js'
+import slowDown from '../../source/index.js'
 
 describe('options', () => {
 	it('should not modify the options object passed', () => {
